# activity-service/api/openapi.yaml
openapi: 3.1.0
info:
  title: FocusNest Activity Service
  version: 0.1.0
servers:
  - url: https://activity.focusnest.com
    description: Production
  - url: http://localhost:8080
    description: Local development
paths:
...
  /v1/users/profile:
    get:
      operationId: getUserProfile
      summary: Get user profile
      tags:
        - user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createUserProfile
      summary: Create user profile
      tags:
        - user
      security:
        - bearerAuth: []
      responses:
        "201":
          description: User profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Profile already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateUserProfile
      summary: Update user profile
      tags:
        - user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileRequest"
      responses:
        "200":
          description: User profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUserProfile
      summary: Delete user profile
      tags:
        - user
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Profile deleted successfully
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
        - code
        - message
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        bio:
          type: string
        birthdate:
          type: string
          format: date
        backgroundImage:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - createdAt
        - updatedAt
    UpdateUserProfileRequest:
      type: object
      properties:
        bio:
          type: string
        birthdate:
          type: string
          format: date
        backgroundImage:
          type: string